requirements.txt
django
djangorestframework
python-dotenv
dnspython
python-whois
python-nmap
requests
openai

